Actualizar services.dart y add_glucose_level.dart

+++ lib/user_data/controller/services.dart
@@ -53,7 +53,7 @@ class UserDataServices {
     return entities;
   }
 
-  Future<void> saveGlucoseLevel(double level) async {
+  Future<void> saveGlucoseLevel(int level) async {
     await _backend.initialize();
     String url = '/api/v1/glucose-levels/';
     await _backend.post(url, {'level': level});
